requirements.txt

# Core Framework
flask>=2.3.3
flask-sqlalchemy>=3.1.1
flask-cors>=4.0.0

# Database
PyMySQL>=1.1.0

# Environment & Configuration
python-dotenv>=1.0.0

# Authentication & Security
PyJWT>=2.8.0
bcrypt>=4.1.0

# API Documentation
flasgger>=0.9.7
flask-swagger-ui>=4.11.1

# Development Tools
black>=23.11.0
pylint>=3.0.2
pytest>=7.4.3
